code/data_storage/python/data/requirements.txt
# SPDX-License-Identifier: GPL-3.0-or-later
# Copyright (C) 2025-2026 ggeoffre, LLC

cassandra-driver
pymongo
pymysql
psycopg2-binary
redis
